feat: add onboarding for compare prices and stores near me

+++ src/components/sale/index.jsx
@@ -568,13 +568,17 @@ const Index = () => {
     //setSelectedLocation(null);
   };
 
+  
   useEffect(() => {
     const handleBeforeUnload = () => {
-      sessionStorage.clear();
+      // Проверяем, был ли переход или именно обновление
+      if (performance.getEntriesByType("navigation")[0]?.type === "reload") {
+        sessionStorage.clear();
+      }
     };
-
+  
     window.addEventListener("beforeunload", handleBeforeUnload);
-
+  
     return () => {
       window.removeEventListener("beforeunload", handleBeforeUnload);
     };
